Moved common filters to one place
